Fix #55, Moves interface definition files to inc
