+++ fsw/inc/lc_platform_cfg.h
