+++ fsw/tables/lc_def_adt.c
@@ -38,8 +38,8 @@
 #include "cfe_tbl_filedef.h"
 #include "lc_platform_cfg.h"
 #include "lc_msgdefs.h"
-#include "lc_tbl.h"
-#include "lc_app.h"
+#include "lc_extern_typedefs.h"
+#include "lc_tbldefs.h"
 #include "lc_events.h"
 
 /*************************************************************************
