+++ fsw/tables/lc_def_wdt.c
@@ -34,8 +34,8 @@
 #include "cfe_tbl_filedef.h"
 #include "lc_platform_cfg.h"
 #include "lc_msgdefs.h"
-#include "lc_tbl.h"
-#include "lc_app.h"
+#include "lc_extern_typedefs.h"
+#include "lc_tbldefs.h"
 
 /*************************************************************************
 ** Examples
